Fixed move

+++ lib/Fuse4Win/fusemain.cpp
@@ -6,6 +6,8 @@
 #include "fuse_win.h"
 #include "utils.h"
 
+#include <stdio.h>
+
 ///////////////////////////////////////////////////////////////////////////////////////
 ////// FUSE bridge
 ///////////////////////////////////////////////////////////////////////////////////////
@@ -532,8 +534,16 @@ int impl_fuse_context::move_file(LPCWSTR file_name, LPCWSTR new_file_name,
 	if (ops_.getattr(new_name.c_str(),&stbuf)!=-ENOENT)
 	{
 		//Cannot delete directory
-		if ((stbuf.st_mode&S_IFDIR)==0 && ops_.unlink){
-            CHECKED(ops_.unlink(new_name.c_str()));
+		if ((stbuf.st_mode&S_IFDIR)==0 && ops_.unlink)
+        {
+            if (replace_existing)
+            {
+                CHECKED(ops_.unlink(new_name.c_str()));
+            }
+            else
+            {
+                return -EEXIST;
+            }
 		}
 	}
